move to a class
